+++ packages/charting/src/components/LineChart/LineChart.types.ts
@@ -2,7 +2,7 @@ import { ITheme, IStyle } from 'office-ui-fabric-react/lib/Styling';
 import { IStyleFunctionOrObject } from 'office-ui-fabric-react/lib/Utilities';
 import { IChartProps } from '../../types/IDataPoint';
 
-export { IChartProps, IDataPoint, ILineChartPoints } from '../../types/IDataPoint';
+export { IChartProps, IDataPoint, ILineChartDataPoint, ILineChartPoints } from '../../types/IDataPoint';
 
 export interface ILineChart {}
 
@@ -41,6 +41,19 @@ export interface ILineChartProps {
    * Width of line stroke
    */
   strokeWidth?: number;
+
+  /**
+   * this prop takes values that you want the line chart to render on x-axis
+   * This is a optional parameter if not specified D3 will decide which values appear on the x-axis for you
+   * Please look at https://github.com/d3/d3-scale for more information on how D3 decides what data to appear on the axis of chart
+   */
+  tickValues?: number[] | Date[];
+
+  /**
+   * the format in for the data on x-axis. For date object this can be specified to your requirement. Eg: '%m/%d', '%d'
+   * Please look at https://www.npmjs.com/package/d3-time-format for all the formats supported
+   */
+  tickFormat?: string;
 }
 
 export interface ILineChartStyleProps {