+++ packages/fluent-theme/src/fluent/styles/PrimaryButton.styles.ts
@@ -1,0 +1,7 @@
+import { fluentBorderRadius } from './styleConstants';
+
+export const PrimaryButtonStyles = {
+  root: {
+    borderRadius: fluentBorderRadius
+  }
+};