+++ packages/utilities/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @uifabric/utilities
 
-This log was last generated on Mon, 15 Oct 2018 12:29:12 GMT and should not be manually modified.
+This log was last generated on Thu, 18 Oct 2018 20:22:36 GMT and should not be manually modified.
+
+## 6.23.1
+Thu, 18 Oct 2018 20:22:36 GMT
+
+### Patches
+
+- Update api-extractor.json
 
 ## 6.23.0
 Mon, 15 Oct 2018 12:29:12 GMT